@@ -1,6 +1,6 @@
 {
   "name": "chatbot-back",
-  "version": "1.7.1",
+  "version": "1.8.0",
   "description": "",
   "author": "",
   "license": "MIT",
@@ -27,72 +27,72 @@
   },
   "dependencies": {
     "@nestjs-modules/mailer": "^1.8.1",
-    "@nestjs/axios": "^0.1.0",
-    "@nestjs/common": "^8.4.7",
+    "@nestjs/axios": "^1.0.1",
+    "@nestjs/common": "^9.2.1",
     "@nestjs/config": "^2.2.0",
-    "@nestjs/core": "^8.4.7",
-    "@nestjs/jwt": "^8.0.1",
-    "@nestjs/passport": "^8.2.2",
-    "@nestjs/platform-express": "^8.4.7",
+    "@nestjs/core": "^9.2.1",
+    "@nestjs/jwt": "^9.0.0",
+    "@nestjs/passport": "^9.0.0",
+    "@nestjs/platform-express": "^9.2.1",
     "@nestjs/schedule": "^2.1.0",
-    "@nestjs/swagger": "^5.1.5",
-    "@nestjs/terminus": "^8.1.1",
-    "@nestjs/typeorm": "~8.0.2",
+    "@nestjs/swagger": "^6.1.4",
+    "@nestjs/terminus": "^9.1.4",
+    "@nestjs/typeorm": "^9.0.1",
     "ansible-playbook-cli-js": "^2.0.5",
-    "archiver": "^4.0.1",
-    "bcrypt": "^5.0.1",
-    "camelcase-keys": "^6.2.1",
-    "class-transformer": "^0.3.1",
-    "class-validator": "^0.13.2",
+    "archiver": "^5.3.1",
+    "bcrypt": "^5.1.0",
+    "camelcase-keys": "^7.0.2",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.0",
     "compression": "^1.7.4",
-    "dotenv": "^8.2.0",
-    "express-rate-limit": "^5.1.1",
+    "dotenv": "^16.0.3",
+    "express-rate-limit": "^6.7.0",
     "get-folder-size": "^2.0.1",
     "js-yaml": "^4.1.0",
-    "mkdirp": "^1.0.4",
-    "moment": "^2.29.3",
-    "nestjs-typeorm-paginate": "^3.2.1",
-    "nodemailer": "^6.6.1",
+    "mkdirp": "^2.0.0",
+    "moment": "^2.29.4",
+    "nestjs-typeorm-paginate": "^4.0.3",
+    "nodemailer": "^6.9.0",
     "passport": "^0.6.0",
     "passport-headerapikey": "^1.2.2",
-    "passport-jwt": "^4.0.0",
+    "passport-jwt": "^4.0.1",
     "passport-local": "^1.0.0",
     "pg": "^8.8.0",
     "pg-escape": "^0.2.0",
     "reflect-metadata": "^0.1.13",
-    "rimraf": "^3.0.0",
-    "rxjs": "^7.1.0",
-    "snakecase-keys": "^3.1.2",
-    "swagger-ui-express": "^4.3.0",
-    "typeorm": "^0.2.38",
-    "uuid": "^7.0.3",
-    "xlsx": "^0.17.2"
+    "rimraf": "^4.0.7",
+    "rxjs": "^7.8.0",
+    "snakecase-keys": "^5.4.4",
+    "swagger-ui-express": "^4.6.0",
+    "typeorm": "^0.3.11",
+    "uuid": "^9.0.0",
+    "xlsx": "^0.18.5"
   },
   "devDependencies": {
-    "@nestjs/cli": "^8.2.8",
-    "@nestjs/schematics": "^8.0.11",
-    "@nestjs/testing": "^8.4.7",
-    "@types/express": "^4.17.2",
-    "@types/jest": "26.0.4",
-    "@types/node": "^13.1.6",
-    "@types/passport-jwt": "^3.0.3",
-    "@types/passport-local": "^1.0.33",
-    "@types/supertest": "^2.0.8",
-    "@typescript-eslint/eslint-plugin": "^2.12.0",
-    "@typescript-eslint/parser": "^2.12.0",
-    "artillery": "^1.7.9",
+    "@nestjs/cli": "^9.1.8",
+    "@nestjs/schematics": "^9.0.4",
+    "@nestjs/testing": "^9.2.1",
+    "@types/express": "^4.17.15",
+    "@types/jest": "29.2.5",
+    "@types/node": "^18.11.18",
+    "@types/passport-jwt": "^3.0.8",
+    "@types/passport-local": "^1.0.35",
+    "@types/supertest": "^2.0.12",
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.1",
+    "artillery": "^2.0.0-27",
     "artillery-engine-socketio-v3": "^1.1.2",
-    "eslint": "^6.7.2",
-    "eslint-config-prettier": "^6.7.0",
-    "eslint-plugin-import": "^2.19.1",
-    "jest": "^29.0.3",
-    "prettier": "^1.18.2",
-    "supertest": "^4.0.2",
-    "ts-jest": "^29.0.1",
-    "ts-loader": "^6.2.1",
+    "eslint": "^8.32.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-import": "^2.27.4",
+    "jest": "^29.3.1",
+    "prettier": "^2.8.3",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.0.5",
+    "ts-loader": "^9.4.2",
     "ts-node": "^10.9.1",
-    "tsconfig-paths": "^3.9.0",
-    "typescript": "^4.8.3"
+    "tsconfig-paths": "^4.1.2",
+    "typescript": "^4.9.4"
   },
   "jest": {
     "moduleFileExtensions": [
