@@ -2,22 +2,23 @@ import { Body, Controller, Delete, Get, Param, Post, Put, Req, UseGuards } from 
 import { UserService } from "./user.service";
 import { UserDto } from "@core/dto/user.dto";
 import { ApiBearerAuth, ApiBody, ApiOperation, ApiTags } from "@nestjs/swagger";
-import { plainToClass } from "class-transformer";
-import camelcaseKeys = require("camelcase-keys");
+import { plainToInstance } from "class-transformer";
 import { User } from "@core/entities/user.entity";
 import { CreateUserDto } from "@core/dto/create-user.dto";
-import snakecaseKeys = require("snakecase-keys");
 import { UserModel } from "@core/models/user.model";
 import { JwtGuard } from "@core/guards/jwt.guard";
 import { RolesGuard } from "@core/guards/roles.guard";
 import { UserRole } from "@core/enums/user-role.enum";
 import { Roles } from "@core/decorators/roles.decorator";
 import { UpdateUserDto } from "@core/dto/update-user.dto";
+import camelcaseKeys = require("camelcase-keys");
+import snakecaseKeys = require("snakecase-keys");
 
 @Controller('user')
 @ApiTags('user')
 export class UserController {
-  constructor(private readonly _userService: UserService) {}
+  constructor(private readonly _userService: UserService) {
+  }
 
   @Get('')
   @ApiOperation({summary: 'Return all users'})
@@ -34,7 +35,7 @@ export class UserController {
       users = users.filter(u => u.email === userRequest.email);
     }
 
-    return plainToClass(UserDto, camelcaseKeys(users, {deep: true}));
+    return plainToInstance(UserDto, camelcaseKeys(users, {deep: true}));
   }
 
   // TODO: Pour le développement, à voir commencer insérer les users par la suite
@@ -45,8 +46,8 @@ export class UserController {
     type: CreateUserDto,
   })
   async generateAdminUser(@Body() user: CreateUserDto): Promise<UserDto> {
-    const userModel = await this._userService.generateAdminUser(plainToClass(UserModel, snakecaseKeys(user)));
-    return plainToClass(UserDto, camelcaseKeys(userModel, {deep: true}));
+    const userModel = await this._userService.generateAdminUser(plainToInstance(UserModel, snakecaseKeys(user)));
+    return plainToInstance(UserDto, camelcaseKeys(userModel, {deep: true}));
   }
 
   @Post('')
@@ -59,8 +60,8 @@ export class UserController {
   @UseGuards(JwtGuard, RolesGuard)
   @Roles(UserRole.admin)
   async create(@Body() user: CreateUserDto): Promise<UserDto> {
-    const userModel = await this._userService.create(plainToClass(UserModel, snakecaseKeys(user)));
-    return plainToClass(UserDto, camelcaseKeys(userModel, {deep: true}));
+    const userModel = await this._userService.create(plainToInstance(UserModel, snakecaseKeys(user)));
+    return plainToInstance(UserDto, camelcaseKeys(userModel, {deep: true}));
   }
 
   @Put(':email')
@@ -74,8 +75,8 @@ export class UserController {
   @Roles(UserRole.admin)
   async update(@Body() user: UpdateUserDto,
                @Param('email') email: string): Promise<UserDto> {
-    const userModel = await this._userService.update(email, plainToClass(UserModel, snakecaseKeys(user)));
-    return plainToClass(UserDto, camelcaseKeys(userModel, {deep: true}));
+    const userModel = await this._userService.update(email, plainToInstance(UserModel, snakecaseKeys(user)));
+    return plainToInstance(UserDto, camelcaseKeys(userModel, {deep: true}));
   }
 
   @Delete(':email')
