@@ -3,7 +3,7 @@ import { ResponseService } from "./response.service";
 import { ApiBearerAuth, ApiOperation, ApiTags } from "@nestjs/swagger";
 import { JwtGuard } from "@core/guards/jwt.guard";
 import { ResponseDto } from "@core/dto/response.dto";
-import { plainToClass } from "class-transformer";
+import { plainToInstance } from "class-transformer";
 import { Response } from "@core/entities/response.entity";
 import camelcaseKeys = require("camelcase-keys");
 import { ResponseModel } from "@core/models/response.model";
@@ -23,19 +23,19 @@ export class ResponseController {
   @ApiOperation({ summary: 'Return all responses' })
   async getReponses(): Promise<ResponseDto[]> {
     const responses: Response[] = await this._responseService.findAll();
-    return plainToClass(ResponseDto, camelcaseKeys(responses, {deep: true}));
+    return plainToInstance(ResponseDto, camelcaseKeys(responses, {deep: true}));
   }
 
   @Post('')
   @ApiOperation({summary: 'Create a response'})
   async createResponse(@Body() response: ResponseDto): Promise<ResponseDto> {
-    const responseToReturn: Response = await this._responseService.create(plainToClass(ResponseModel, snakecaseKeys(response)));
-    return plainToClass(ResponseDto, camelcaseKeys(responseToReturn, {deep: true}));
+    const responseToReturn: Response = await this._responseService.create(plainToInstance(ResponseModel, snakecaseKeys(response)));
+    return plainToInstance(ResponseDto, camelcaseKeys(responseToReturn, {deep: true}));
   }
 
   @Put(':id')
   @ApiOperation({summary: 'Edit a response'})
   async editResponse(@Param('id') responseId: string, @Body() response: UpdateResponseDto): Promise<UpdateResult> {
-    return this._responseService.update(plainToClass(ResponseModel, snakecaseKeys(response)));
+    return this._responseService.update(plainToInstance(ResponseModel, snakecaseKeys(response)));
   }
 }
